+++ ruleminer/parser.py
@@ -46,6 +46,7 @@ class RuleParser:
             (set(["contains", "not contains"]), self.parse_contains),
             (set(["exact"]), self.parse_exact),
             (set(["corr"]), self.parse_corr),
+            (set(["table"]), self.parse_table),
             (set(["abs"]), self.parse_abs),
             (set(["max", "min", "abs"]), self.parse_maxminabs),
             (set(["round", "floor", "ceil"]), self.parse_round),
@@ -281,6 +282,48 @@ class RuleParser:
             corr_params,
             apply_tolerance=apply_tolerance,
             positive_tolerance=positive_tolerance,
+        )
+        return res
+
+    def parse_table(
+        self,
+        idx: int,
+        item: str,
+        expression: Union[str, list],
+        apply_tolerance: bool = False,
+        positive_tolerance: bool = True,
+    ) -> str:
+        """
+        Process corr function
+
+        Example:
+            expression = ['TABLE', ['(', '"table_name"', ',', '{"a"}', ',', '{"b"}', ',', '{"c"}, ',', '{"d"}',')']]
+            idx = 0
+
+            result = ruleminer.RuleParser().parse_corr(
+                idx=idx,
+                expression=expression,
+                apply_tolerance=False
+            )
+            print(result)
+                '__table_name({"a"}, {"b"}, {"c"}, {"d"})'
+        """
+        table_params = expression[idx + 1][3:][0]
+        table_name = expression[idx + 1][1][1:-1]
+        if table_name not in list(self.params.get("tables", {}).keys()):
+            logging.error(
+                "Table name is not in predefined tables dictionary of parameters."
+            )
+        res = (
+            "_table_"
+            + table_name
+            + "["
+            + self.parse(
+                table_params,
+                apply_tolerance=apply_tolerance,
+                positive_tolerance=positive_tolerance,
+            )
+            + "].apply(tuple, axis=1)"
         )
         return res
 
@@ -738,6 +781,14 @@ class RuleParser:
                 apply_tolerance=apply_tolerance,
                 positive_tolerance=positive_tolerance,
             )
+        if left_side[0][0] == "[":
+            # if the left side is a list of columns then we concat the columns to a dataframe and convert it to tuples
+            left_side_list = self.parse(
+                left_side[0],
+                apply_tolerance=apply_tolerance,
+                positive_tolerance=positive_tolerance,
+            )
+            res = "pd.concat(" + left_side_list + ", axis=1).apply(tuple, axis=1)"
         if item.lower() in ["in", "not in"]:
             res += ".isin("
             for i in right_side: